Forcing the user to comply with alternative command input requested
Thanks to user http://emacs.stackexchange.com/users/9198/compunaut for
helping me figuring an optimal solution for problem as posted here:
http://emacs.stackexchange.com/q/27780/690

The description of the package was modified. More features were added
too. The counter will not work inside the minibuffer, this avoid the
alternative command to be execute inside of it.

If the user decides to exit the alternative command with C-g,
post-command-hook will never see the `quit` error.

The minimum Emacs version is the release one released in March 26, 2008,
because subr.el's string-match-p was implemented by RMS one year
early... but I can be wrong an be using a feature from subr.el not
available to Emacs 22.2.