chg: Advises the user for any function that cause recursion errors
Some functions like `next-line` throws an error if they reach the end of the buffer, which is good,
however, this can make schrute-mode to enter into an infinite recursion because it thinks the
alternative function hasn't finished correctly or the user didn't comply with any required input
causing the major mode to fail miserably lol.

We still have a bug that is caused when the user keeps invoking the same command and the alternative
commands works without requesting input, somehow causing the recursion error I talked early. I still
need to debug that.

refs bug #1